refactor: move the agent terminal panel into components/
The media sharing page needs the same Claude Code panel the homepage uses, so
it stops belonging to app/home/agent-ready/. Sizing and entrance move to a
className prop: the panel owns how it looks, the caller owns how big it is.
